third_party/WebKit/LayoutTests/platform/win7/external/wpt/mediacapture-streams/MediaStreamTrack-getSettings.https-expected.txt

This is a testharness.js-based test.
FAIL A device can be opened twice and have the same device ID assert_equals: expected "ba74ff53-f6f5-4f8f-af52-e0ea3960274b" but got "b3955035-79eb-43d5-94f2-8058b777bb91"
FAIL A device can be opened twice with different resolutions assert_equals: expected "7909db79-5833-4e86-b8b6-ee991923a0ab" but got "b57f4533-39c8-45da-bb10-b13bca52dce8"
Harness: the test ran to completion.
